requirements.txt
wandb==0.12.15
cityscapesscripts==2.2.0
scipy==1.8.0
opencv-python==4.5.5.64
seaborn==0.11.2
pillow==9.1.0
streamlit==1.8.1
torch==1.11.0
torchvision==0.12.0
typer==0.4.1
matplotlib==3.5.1
scikit-image==3.0.8
filterpy==1.4.5